requirements.txt
# Core Dependencies
streamlit>=1.30.0
python-dotenv>=1.0.0
openai>=1.3.0
pyautogen>=0.2.0

# Code Analysis & Quality
pylint>=3.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Utilities
pyyaml>=6.0
requests>=2.31.0
aiohttp>=3.9.0

# CI/CD & DevOps
docker>=6.1.0
gitpython>=3.1.40

# Security
bandit>=1.7.5
safety>=2.3.5

# Performance Monitoring
psutil>=5.9.0

# Data Handling
pandas>=2.1.0
numpy>=1.24.0
